+++ src/components/mod.rs
@@ -1,6 +1,7 @@
 mod button;
 mod icon;
 mod icon_button;
+mod radio_button;
 mod state_layer;
 mod surface;
 mod typography;
@@ -9,6 +10,7 @@ pub use self::{
     button::{Button, ButtonStyle},
     icon::{Icon, IconData, IconStyle},
     icon_button::{IconButton, IconButtonStyle},
+    radio_button::RadioButton,
     state_layer::StateLayer,
     surface::Surface,
     typography::Typography,