+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `component:IconButton`
   - `component:Button`
   - `component:Icon`
+  - `component:RadioButton`
   - `component:StateLayer`
   - `component:Typography`